// sarif/Sarif.d.ts
export interface Sarif {
  /**
   * The URI of the JSON schema corresponding to the version.
   */
  $schema?: string;
  /**
   * The SARIF format version of this log file.
   */
  version: "1.0.0";
  /**
   * The set of runs contained in this log file.
   */
  runs: Run[];
}
// ...
export interface File {
  /**
   * The path to the file within its containing file.
   */
  uri?: string;
  /**
   * A string that identifies the conceptual base for the 'uri' property (if it is relative), e.g.,'$(SolutionDir)' or '%SRCROOT%'.
   */
  uriBaseId?: string;
  /**
   * Identifies the key of the immediate parent of the file, if this file is nested.
   */
  parentKey?: string;
  /**
   * The offset in bytes of the file within its containing file.
   */
  offset?: number;
  /**
   * The length of the file in bytes.
   */
  length?: number;
  /**
   * The MIME type (RFC 2045) of the file.
   */
  mimeType?: string;
  /**
   * The contents of the file, expressed as a MIME Base64-encoded byte sequence.
   */
  contents?: string;
  /**
   * An array of hash objects, each of which specifies a hashed value for the file, along with the name of the algorithm used to compute the hash.
   */
  hashes?: Hash[];
  /**
   * Key/value pairs that provide additional information about the file.
   */
  properties?: {
    /**
     * A set of distinct strings that provide additional information.
     */
    tags?: string[];
    [k: string]: any;
  };
}
// ...
export interface Result {
  /**
   * The stable, unique identifier of the rule (if any) to which this notification is relevant. If 'ruleKey' is not specified, this member can be used to retrieve rule metadata from the rules dictionary, if it exists.
   */
  ruleId?: string;
  /**
   * A key used to retrieve the rule metadata from the rules dictionary that is relevant to the notificationn.
   */
  ruleKey?: string;
  /**
   * A value specifying the severity level of the result. If this property is not present, its implied value is 'warning'.
   */
  level?: "notApplicable" | "pass" | "note" | "warning" | "error";
  /**
   * A string that describes the result. The first sentence of the message only will be displayed when visible space is limited.
   */
  message?: string;
  /**
   * A 'formattedRuleMessage' object that can be used to construct a formatted message that describes the result. If the 'formattedMessage' property is present on a result, the 'fullMessage' property shall not be present. If the 'fullMessage' property is present on an result, the 'formattedMessage' property shall not be present
   */
  formattedRuleMessage?: {
    /**
     * A string that identifies the message format used to format the message that describes this result. The value of formatId must correspond to one of the names in the set of name/value pairs contained in the 'messageFormats' property of the rule object whose 'id' property matches the 'ruleId' property of this result.
     */
    formatId: string;
    /**
     * An array of strings that will be used, in combination with a message format, to construct a result message.
     */
    arguments?: string[];
  };
  /**
   * One or more locations where the result occurred. Specify only one location unless the problem indicated by the result can only be corrected by making a change at every specified location.
   */
  locations?: Location[];
  /**
   * A source code or other file fragment that illustrates the result.
   */
  snippet?: string;
  /**
   * A unique identifer for the result.
   */
  id?: string;
  /**
   * A string that contributes to the unique identity of the result.
   */
  toolFingerprintContribution?: string;
  /**
   * An array of 'stack' objects relevant to the result.
   */
  stacks?: Stack[];
  /**
   * An array of 'codeFlow' objects relevant to the result.
   */
  codeFlows?: CodeFlow[];
  /**
   * A grouped set of locations and messages, if available, that represent code areas that are related to this result.
   */
  relatedLocations?: AnnotatedCodeLocation[];
  suppressionStates?: ("suppressedInSource" | "suppressedExternally")[];
  /**
   * The state of a result relative to a baseline of a previous run.
   */
  baselineState?: "new" | "existing" | "absent";
  /**
   * An array of 'fix' objects, each of which represents a proposed fix to the problem indicated by the result.
   */
  fixes?: Fix[];
  /**
   * Key/value pairs that provide additional information about the result.
   */
  properties?: {
    /**
     * A set of distinct strings that provide additional information.
     */
    tags?: string[];
    [k: string]: any;
  };
}
// ...
export interface Stack {
  /**
   * A message relevant to this call stack.
   */
  message?: string;
  /**
   * An array of stack frames that represent a sequence of calls, rendered in reverse chronological order, that comprise the call stack.
   */
  frames: StackFrame[];
  /**
   * Key/value pairs that provide additional information about the stack.
   */
  properties?: {
    /**
     * A set of distinct strings that provide additional information.
     */
    tags?: string[];
    [k: string]: any;
  };
}
// ...
export interface CodeFlow {
  /**
   * A message relevant to the code flow
   */
  message?: string;
  /**
   * An array of 'annotatedCodeLocation' objects, each of which describes a single location visited by the tool in the course of producing the result.
   */
  locations: AnnotatedCodeLocation[];
  /**
   * Key/value pairs that provide additional information about the code flow.
   */
  properties?: {
    /**
     * A set of distinct strings that provide additional information.
     */
    tags?: string[];
    [k: string]: any;
  };
  [k: string]: any;
}
/**
 * An annotation used to express code flows through a method or other locations that are related to a result.
 */
export interface AnnotatedCodeLocation {
  /**
   * OBSOLETE (use "step" instead): An identifier for the location, unique within the scope of the code flow within which it occurs.
   */
  id?: number | string;
  /**
   * The 0-based sequence number of the location in the code flow within which it occurs.
   */
  step?: number;
  /**
   * A file location to which this annotation refers.
   */
  physicalLocation?: {
    /**
     * The location of the file as a valid URI.
     */
    uri?: string;
    /**
     * A string that identifies the conceptual base for the 'uri' property (if it is relative), e.g.,'$(SolutionDir)' or '%SRCROOT%'.
     */
    uriBaseId?: string;
    /**
     * The region within the file where the result was detected.
     */
    region?: {
      /**
       * The line number of the first character in the region.
       */
      startLine?: number;
      /**
       * The column number of the first character in the region.
       */
      startColumn?: number;
      /**
       * The line number of the last character in the region.
       */
      endLine?: number;
      /**
       * The column number of the last character in the region.
       */
      endColumn?: number;
      /**
       * The zero-based offset from the beginning of the file of the first byte or character in the region.
       */
      offset?: number;
      /**
       * The length of the region in bytes or characters.
       */
      length?: number;
    };
  };
  /**
   * The fully qualified name of the method or function that is executing.
   */
  fullyQualifiedLogicalName?: string;
  /**
   * A key used to retrieve the annotation's logicalLocation from the logicalLocations dictionary.
   */
  logicalLocationKey?: string;
  /**
   * The name of the module that contains the code that is executing.
   */
  module?: string;
  /**
   * The thread identifier of the code that is executing.
   */
  threadId?: number;
  /**
   * A message relevant to this annotation.
   */
  message?: string;
  /**
   * Categorizes the location.
   */
  kind?:
    | "alias"
    | "assignment"
    | "branch"
    | "call"
    | "callReturn"
    | "continuation"
    | "declaration"
    | "functionEnter"
    | "functionExit"
    | "functionReturn"
    | "usage";
  /**
   * Classifies state transitions in code locations relevant to a taint analysis.
   */
  taintKind?: "source" | "sink" | "sanitizer";
  /**
   * The fully qualified name of the target on which this location operates. For an annotation of kind 'call', for example, the target refers to the fully qualified logical name of the function called from this location.
   */
  target?: string;
  /**
   * An ordered set of strings that comprise input or return values for the current operation. For an annotation of kind 'call', for example, this property may hold the ordered list of arguments passed to the callee.
   */
  values?: string[];
  /**
   * A dictionary, each of whose keys specifies a variable or expression, the associated value of which represents the variable or expression value. For an annotation of kind 'continuation', for example, this dictionary might hold the current assumed values of a set of global variables.
   */
  state?: {
    [k: string]: any;
  };
  /**
   * A key used to retrieve the target's logicalLocation from the logicalLocations dictionary.
   */
  targetKey?: string;
  /**
   * OBSOLETE (use "importance" instead): True if this location is essential to understanding the code flow in which it occurs.
   */
  essential?: boolean;
  /**
   * Specifies the importance of this location in understanding the code flow in which it occurs. The order from most to least important is "essential", "important", "unimportant". Default: "important".
   */
  importance?: "important" | "essential" | "unimportant";
  /**
   * The source code at the specified location.
   */
  snippet?: string;
  /**
   * A set of messages relevant to the current annotated code location.
   */
  annotations?: Annotation[];
  /**
   * Key/value pairs that provide additional information about the code location.
   */
  properties?: {
    /**
     * A set of distinct strings that provide additional information.
     */
    tags?: string[];
    [k: string]: any;
  };
}
export interface Annotation {
  /**
   * A message relevant to a code location
   */
  message: string;
  /**
   * An array of 'physicalLocation' objects associated with the annotation.
   */
  locations: PhysicalLocation[];
  [k: string]: any;
}
// ...
export interface Fix {
  /**
   * A string that describes the proposed fix, enabling viewers to present a proposed change to an end user.
   */
  description?: string;
  /**
   * One or more file changes that comprise a fix for a result.
   */
  fileChanges: FileChange[];
}
// ...
export interface Rule {
  /**
   * A stable, opaque identifier for the rule.
   */
  id: string;
  /**
   * A rule identifier that is understandable to an end user.
   */
  name?: string;
  /**
   * A concise description of the rule. Should be a single sentence that is understandable when visible space is limited to a single line of text.
   */
  shortDescription?: string;
  /**
   * A string that describes the rule. Should, as far as possible, provide details sufficient to enable resolution of any problem indicated by the result.
   */
  fullDescription?: string;
  /**
   * A set of name/value pairs with arbitrary names. The value within each name/value pair shall consist of plain text interspersed with placeholders, which can be used to format a message in combination with an arbitrary number of additional string arguments.
   */
  messageFormats?: {
    [k: string]: any;
  };
  /**
   * A value specifying whether a rule is enabled.
   */
  configuration?: "unknown" | "enabled" | "disabled";
  /**
   * A value specifying the default severity level of the result.
   */
  defaultLevel?: "note" | "warning" | "error";
  /**
   * A URI where the primary documentation for the rule can be found.
   */
  helpUri?: string;
  /**
   * Key/value pairs that provide additional information about the rule.
   */
  properties?: {
    /**
     * A set of distinct strings that provide additional information.
     */
    tags?: string[];
    [k: string]: any;
  };
}